+++ src/hkimjp/konpy2/answers.clj
@@ -62,9 +62,10 @@
   (let [author (user request)
         answer (slurp (:tempfile file))
         e (parse-long e)
-        test-code (:test (ds/pl e))]
+        testcode (:testcode (ds/pl e))]
+    (t/log! {:level :debug :data {:testcode testcode}})
     (try
-      (validate author answer test-code)
+      (validate author answer testcode)
       (ds/put! {:answer/status "yes"
                 :to      e
                 :author  author
@@ -73,7 +74,7 @@
                 :updated (now)})
       (redirect (str "/k/problem/" e))
       (catch Exception ex
-        (t/log! {:level :error :data file})
+        (t/log! {:level :error :data answer})
         (page
          [:div
           [:div.text-2xl.text-red-600 "Error"]